fixing internal libs from int64 to int

+++ stdlib/rand/rand.go
@@ -6,13 +6,13 @@ import (
 )
 
 //RandomInt Generates a random number between min and max.
-func RandomInt(min, max int64) int {
+func RandomInt(min, max int) int {
 	r, _ := random.Int(random.Reader, big.NewInt(int64(max-min)))
-	return int(r.Int64() + min)
+	return int(r.Int64() + int64(min))
 }
 
 //GetAlphaNumericString Generates a random alpha numeric string of a specified length
-func GetAlphaNumericString(strlen int64) string {
+func GetAlphaNumericString(strlen int) string {
 	const chars = "0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
 	result := make([]byte, strlen)
 	for i := range result {
@@ -26,7 +26,7 @@ func GetAlphaNumericString(strlen int64) string {
 }
 
 //GetAlphaString generates a random alpha string of a specified length
-func GetAlphaString(strlen int64) string {
+func GetAlphaString(strlen int) string {
 	const chars = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
 	result := make([]byte, strlen)
 	for i := range result {
@@ -40,7 +40,7 @@ func GetAlphaString(strlen int64) string {
 }
 
 //GetAlphaNumericSpecialString generates a random alpha numeric and special char string of a specified length
-func GetAlphaNumericSpecialString(strlen int64) string {
+func GetAlphaNumericSpecialString(strlen int) string {
 	const chars = "0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ!@#$%^&*()"
 	result := make([]byte, strlen)
 	for i := range result {
